+++ docs/src/pages/components/rating/CustomizedRating.tsx
@@ -1,14 +1,9 @@
 import * as React from 'react';
 import { experimentalStyled as styled } from '@material-ui/core/styles';
 import Box from '@material-ui/core/Box';
-import Rating, { IconContainerProps } from '@material-ui/core/Rating';
+import Rating from '@material-ui/core/Rating';
 import FavoriteIcon from '@material-ui/icons/Favorite';
 import FavoriteBorderIcon from '@material-ui/icons/FavoriteBorder';
-import SentimentVeryDissatisfiedIcon from '@material-ui/icons/SentimentVeryDissatisfied';
-import SentimentDissatisfiedIcon from '@material-ui/icons/SentimentDissatisfied';
-import SentimentSatisfiedIcon from '@material-ui/icons/SentimentSatisfied';
-import SentimentSatisfiedAltIcon from '@material-ui/icons/SentimentSatisfiedAltOutlined';
-import SentimentVerySatisfiedIcon from '@material-ui/icons/SentimentVerySatisfied';
 import Typography from '@material-ui/core/Typography';
 
 const StyledRating = styled(Rating)({
@@ -19,39 +14,6 @@ const StyledRating = styled(Rating)({
     color: '#ff3d47',
   },
 });
-
-const customIcons: {
-  [index: string]: {
-    icon: React.ReactElement;
-    label: string;
-  };
-} = {
-  1: {
-    icon: <SentimentVeryDissatisfiedIcon />,
-    label: 'Very Dissatisfied',
-  },
-  2: {
-    icon: <SentimentDissatisfiedIcon />,
-    label: 'Dissatisfied',
-  },
-  3: {
-    icon: <SentimentSatisfiedIcon />,
-    label: 'Neutral',
-  },
-  4: {
-    icon: <SentimentSatisfiedAltIcon />,
-    label: 'Satisfied',
-  },
-  5: {
-    icon: <SentimentVerySatisfiedIcon />,
-    label: 'Very Satisfied',
-  },
-};
-
-function IconContainer(props: IconContainerProps) {
-  const { value, ...other } = props;
-  return <span {...other}>{customIcons[value].icon}</span>;
-}
 
 export default function CustomizedRating() {
   return (
@@ -71,13 +33,6 @@ export default function CustomizedRating() {
       />
       <Typography component="legend">10 stars</Typography>
       <Rating name="customized-10" defaultValue={2} max={10} />
-      <Typography component="legend">Custom icon set</Typography>
-      <Rating
-        name="customized-icons"
-        defaultValue={2}
-        getLabelText={(value: number) => customIcons[value].label}
-        IconContainerComponent={IconContainer}
-      />
     </Box>
   );
 }
